PoC: formatter caching as a service wrapper
I'm not 100% sure about this -- as it stands, it's concentrating more
complexity into the AnnotationJSONPresentationService, which is already
responsible for threading a lot of objects together -- but it reduces
both the formatter logic and the service wrapper to manageably simple
objects.

One difference here is that the preloaded version of the service now
doesn't support fall-through for individual lookups, so this will only
work if the preloaded IDs are the *only* IDs used by the formatter. In
the case we're currently using them, I think this holds, but there may
be a case I haven't thought of. I don't think the fall-through would be
too difficult to implement, but would add some complexity to the
wrapper.